fix delete dialog

+++ src/app/components/animals/animals.component.ts
@@ -76,12 +76,12 @@ export class AnimalsComponent {
   }
 
   delete(id: string) {
-    const dataToDelete = doc(this.firestore, 'animals', id);
-    deleteDoc(dataToDelete)
+    this.dialogService.showConfirm('Informacja', 'Dane usunięte').subscribe(result => {
+      const dataToDelete = doc(this.firestore, 'animals', id);
+      if(result){
+        deleteDoc(dataToDelete)
       .then(() => {
-        this.dialogService.showInfo('Informacja', 'Dane usunięte').subscribe(result => {
-      console.log(':', result);
-    });
+        
         this.getData()
       })
       .catch((err) => {
@@ -89,7 +89,10 @@ export class AnimalsComponent {
         this.dialogService.showInfo('Informacja',err.message).subscribe(result => {
       console.log(':', result);
       });
-      })
+      }
+    )}
+    });
+    
 
   }
