+++ src/app/components/warehouse/warehouse.component.ts
@@ -75,8 +75,10 @@ export class WarehouseComponent {
   }
 
   delete(id: string) {
-    const dataToDelete = doc(this.firestore, 'warehouse', id);
-    deleteDoc(dataToDelete)
+    this.dialogService.showConfirm('Informacja', 'Dane usunięte').subscribe(result => {
+      const dataToDelete = doc(this.firestore, 'warehouse', id);
+      if(result){
+        deleteDoc(dataToDelete)
       .then(() => {
         this.dialogService.showInfo('Informacja', 'Dane usunięte').subscribe(result => {
       console.log(':', result);
@@ -87,7 +89,11 @@ export class WarehouseComponent {
         this.dialogService.showInfo('Informacja',err.message).subscribe(result => {
       console.log(':', result);
     });
-      })
+      }
+    )}
+    });
+    
+
   }
 
   edit(i: any) {
